Factorized function to _BaseWrapper
