+++ freediscovery/base.py
@@ -9,6 +9,7 @@ import os
 import numpy as np
 from sklearn.externals import joblib
 
+from .text import FeatureVectorizer
 from .exceptions import ModelNotFound
 from .exceptions import (DatasetNotFound, InitException, NotFound, WrongParameter)
 
@@ -41,148 +42,44 @@ class RankerMixin(object):
         return roc_auc_score(y, self.decision_function(X), sample_weight=sample_weight,)
 
 
-class _BaseTextTransformer(object):
-    """Base text transformer
+class _BaseWrapper(object):
+    """Base class for wrappers in FreeDiscovery
 
     Parameters
     ----------
-    cache_dir : str, default='/tmp/'
-        directory where to save temporary and regression files
-    dsid : str
-        load an exising dataset
-    verbose : bool
-        pring progress messages
+    cache_dir : str
+      folder where the model will be saved
+    dsid : str, optional
+      dataset id
+    mid : str, optional
+      model id
     """
+    def __init__(self, cache_dir='/tmp/', dsid=None, mid=None,
+                 dataset_definition=FeatureVectorizer):
 
-    def __init__(self, cache_dir='/tmp/', dsid=None, verbose=False):
-        self.data_dir = None
-        self.verbose = verbose
-        if cache_dir is not None:
-            self.cache_dir = cache_dir = os.path.join(cache_dir, 'ediscovery_cache')
-            if not os.path.exists(cache_dir):
-                os.mkdir(cache_dir)
-        else:
-            self.cache_dir = None
-        self.dsid = dsid
-        if dsid is not None:
-            dsid_dir = os.path.join(self.cache_dir, dsid)
-            if not os.path.exists(dsid_dir):
-                raise DatasetNotFound()
+        if dsid is None and mid is not None:
+            self.dsid = dsid = self.get_dsid(cache_dir, mid)
+            self.mid = mid
+        elif dsid is not None:
+            self.dsid  = dsid
+            self.mid = None
+        elif dsid is None and mid is None:
+            raise WrongParameter('dsid and mid')
+
+        self.fe = dataset_definition(cache_dir=cache_dir, dsid=dsid)
+
+        self.model_dir = os.path.join(self.fe.cache_dir, dsid, self._wrapper_type)
+
+        if not os.path.exists(self.model_dir):
+            os.mkdir(self.model_dir)
+
+        if self.mid is not None:
             pars = self._load_pars()
         else:
-            dsid_dir = None
             pars = None
-        self.dsid_dir = dsid_dir
         self._pars = pars
 
-    @staticmethod
-    def _list_filenames(data_dir, dir_pattern, file_pattern):
-        import re
-        # parse all files in the folder
-        filenames = []
-        for root, subdirs, files in os.walk(data_dir):
-            #print(root, dir_pattern)
-            if re.match(dir_pattern, root):
-                for fname in files:
-                    if re.match(file_pattern, fname):
-                        filenames.append(os.path.normpath(os.path.join(root, fname)))
 
-        # make sure that sorting order is deterministic
-        return sorted(filenames)
-
-    def delete(self):
-        """ Delete the current dataset """
-        import shutil
-        shutil.rmtree(self.dsid_dir, ignore_errors=True)
-
-    def __contains__(self, dsid):
-        """ This is a somewhat non standard call that checks if a dsid
-        exist on disk (in general)"""
-        dsid_dir = os.path.join(self.cache_dir, dsid)
-        return os.path.exists(dsid_dir)
-
-    def __getitem__(self, index):
-        return np.asarray(self._pars['filenames'])[index]
-
-    def load(self, dsid):
-        """ Load a computed features from disk"""
-        if self.cache_dir is None:
-            raise InitException('cache_dir is None: cannot load from cache!')
-        dsid_dir = os.path.join(self.cache_dir, dsid)
-        if not os.path.exists(dsid_dir):
-            raise DatasetNotFound('dsid not found!')
-        pars = joblib.load(os.path.join(dsid_dir, 'pars'))
-        fset_new = joblib.load(os.path.join(dsid_dir, 'features'))
-        return pars['filenames'], fset_new
-
-    def get_params(self):
-        """ Get the vectorizer parameters """
-        return self._pars
-
-    def _load_pars(self):
-        """ Load parameters from disk"""
-        dsid = self.dsid
-        if self.cache_dir is None:
-            raise InitException('cache_dir is None: cannot load from cache!')
-        dsid_dir = os.path.join(self.cache_dir, dsid)
-        if not os.path.exists(dsid_dir):
-            raise DatasetNotFound('dsid {} not found!'.format(dsid))
-        pars = joblib.load(os.path.join(dsid_dir, 'pars'))
-        return pars
-
-    def search(self, filenames):
-        """ Return the document ids that correspond to the provided filenames.
-
-        Parameters
-        ----------
-        filenames : list[str]
-            list of filenames (relatives to the data_dir)
-
-        Returns
-        -------
-        indices : array[int]
-            corresponding list of document id (order is not preserved)
-        """
-        filenames_all = self._pars['filenames']
-        # calculate the indices of the intersection of filenames with filenames_all
-        ind_dict = dict((k,i) for i,k in enumerate(filenames_all))
-        indices = [ ind_dict[x] for x in filenames]
-        return np.array(indices)
-
-    @property
-    def n_samples_(self):
-        """ Number of documents in the dataset """
-        return len(self._pars['filenames'])
-
-    def list_datasets(self):
-        """ List all datasets in the working directory """
-        import traceback
-        out = []
-        for dsid in os.listdir(self.cache_dir):
-            row = {"id": dsid}
-            self.dsid = dsid
-            try:
-                pars = self._load_pars()
-            except:
-                print(pars.keys())
-                traceback.print_exc()
-                continue
-
-            if pars['type'] != type(self).__name__:
-                continue
-
-            try:
-                for key in self._PARS_SHORT:
-                    row[key] = pars[key]
-                out.append(row)
-            except Exception:
-                print(pars.keys())
-                traceback.print_exc()
-
-        return out
-
-
-class _BaseWrapper(object):
     def get_path(self, mid):
         dsid = self.get_dsid(self.fe.cache_dir, mid)
         return os.path.join(self.fe.cache_dir, dsid, self._wrapper_type, mid)
@@ -213,15 +110,30 @@ class _BaseWrapper(object):
         """ Get model parameters """
         return self._pars
 
+    def _load_pars(self, mid=None):
+        """Load model parameters from disk"""
+        if mid is None:
+            mid = self.mid
+        model_path = os.path.join(self.model_dir, mid)
+        if not os.path.exists(model_path):
+            raise ValueError('Model id {} ({}) not found in the cache {}!'.format(
+                             mid, self._wrapper_type, model_path))
+        pars = joblib.load(os.path.join(model_path, 'pars'))
+        pars['id'] = mid
+        return pars
+
     def list_models(self):
+        """ List existing models of this type """
         out = []
+        if not os.path.exists(self.model_dir):
+            return out
         for mid in os.listdir(self.model_dir):
             try:
-                pars = self.load_pars(mid)
+                pars = self._load_pars(mid)
                 pars['id'] = mid
                 out.append(pars)
             except:
-                pass
+                raise
         return out
