+++ freediscovery/categorization.py
@@ -15,7 +15,6 @@ from sklearn.base import BaseEstimator
 from sklearn.utils.validation import check_array
 
 
-from .text import FeatureVectorizer
 from .base import _BaseWrapper, RankerMixin
 from .utils import setup_model, _rename_main_thread
 from .exceptions import (ModelNotFound, WrongParameter, NotImplementedFD, OptionalDependencyMissing)
@@ -225,28 +224,12 @@ class Categorizer(_BaseWrapper):
     def __init__(self, cache_dir='/tmp/',  dsid=None, mid=None,
             cv_scoring='roc_auc', cv_n_folds=3):
 
-        if dsid is None and mid is not None:
-            self.dsid = dsid =  self.get_dsid(cache_dir, mid)
-            self.mid = mid
-        elif dsid is not None:
-            self.dsid  = dsid
-            self.mid = None
-        elif dsid is None and mid is None:
-            raise WrongParameter('dsid and mid')
-
-        self.fe = FeatureVectorizer(cache_dir=cache_dir, dsid=dsid)
-
-        self.model_dir = os.path.join(self.fe.cache_dir, dsid, self._wrapper_type)
-
-        if not os.path.exists(self.model_dir):
-            os.mkdir(self.model_dir)
+        super(Categorizer, self).__init__(cache_dir=cache_dir,  dsid=dsid, mid=mid)
 
         if self.mid is not None:
-            pars, cmod = self._load_pars()
+            cmod = self._load_model()
         else:
-            pars = None
             cmod = None
-        self._pars = pars
         self.cmod = cmod
 
         self.cv_scoring = cv_scoring
@@ -437,13 +420,17 @@ class Categorizer(_BaseWrapper):
         res = np.concatenate(res, axis=0)
         return res
 
-    def _load_pars(self):
-        """ Load the parameters specified by a mid """
-        mid = self.mid
+    def _load_pars(self, mid=None):
+        """Load model parameters from disk"""
+        if mid is None:
+            mid = self.mid
         mid_dir = os.path.join(self.model_dir, mid)
-        if not os.path.exists(mid_dir):
-            raise ModelNotFound('Model id {} not found in the cache!'.format(mid))
-        pars = joblib.load(os.path.join(mid_dir, 'pars'))
+        pars = super(Categorizer, self)._load_pars(mid)
         cmod = joblib.load(os.path.join(mid_dir, 'model'))
         pars['options'] = cmod.get_params()
-        return pars, cmod
+        return pars
+
+    def _load_model(self):
+        """ Load the parameters specified by a mid """
+        mid_dir = os.path.join(self.model_dir, self.mid)
+        return joblib.load(os.path.join(mid_dir, 'model'))
