+++ freediscovery/tests/test_integration.py
@@ -76,7 +76,7 @@ def test_features_hashing(use_hashing, method):
         lsi_id = lsi.mid
         assert lsi.get_dsid(fe.cache_dir, lsi_id) == uuid
         assert lsi.get_path(lsi_id) is not None
-        assert lsi._load_pars(lsi_id) is not None
+        assert lsi._load_pars() is not None
         lsi.load(lsi_id)
 
         idx_gt = lsi.fe.search(ground_truth.index.values)
