+++ freediscovery/tests/test_lsi.py
@@ -36,7 +36,7 @@ def test_lsi():
     assert lsi_res.components_.shape == (n_components, n_features)
     assert lsi.get_dsid(fe.cache_dir, lsi_id) == uuid
     assert lsi.get_path(lsi_id) is not None
-    assert lsi._load_pars(lsi_id) is not None
+    assert lsi._load_pars() is not None
     lsi.load(lsi_id)
 
     idx_gt = lsi.fe.search(ground_truth.index.values)
