+++ freediscovery/threading.py
@@ -40,28 +40,12 @@ class EmailThreading(_BaseWrapper):
     def __init__(self, cache_dir='/tmp/',  dsid=None, mid=None,
                  decode_header=False):
 
-        if dsid is None and mid is not None:
-            self.dsid = dsid =  self.get_dsid(cache_dir, mid)
-            self.mid = mid
-        elif dsid is not None:
-            self.dsid  = dsid
-            self.mid = None
-        elif dsid is None and mid is None:
-            raise WrongParameter('dsid and mid')
-
-        self.fe = EmailParser(cache_dir=cache_dir, dsid=dsid)
-
-        self.model_dir = os.path.join(self.fe.cache_dir, dsid, self._wrapper_type)
-
-        if not os.path.exists(self.model_dir):
-            os.mkdir(self.model_dir)
-
+        super(EmailThreading, self).__init__(cache_dir=cache_dir,  dsid=dsid, mid=mid,
+                                          dataset_definition=EmailParser)
         if self.mid is not None:
-            pars, cmod = self._load_pars()
+            cmod = self._load_model()
         else:
-            pars = None
             cmod = None
-        self._pars = pars
         self.cmod = cmod
 
 
@@ -121,12 +105,11 @@ class EmailThreading(_BaseWrapper):
         self.cmod = cmod
         return cmod
 
-    def _load_pars(self):
-        """ Load the parameters specified by a mid """
+    def _load_model(self):
         mid = self.mid
         mid_dir = os.path.join(self.model_dir, mid)
         if not os.path.exists(mid_dir):
             raise ModelNotFound('Model id {} not found in the cache!'.format(mid))
-        pars = joblib.load(os.path.join(mid_dir, 'pars'))
         cmod = joblib.load(os.path.join(mid_dir, 'model'))
-        return pars, cmod
+        return cmod
+
